Shell Script for compiling static library calculator

+++ core-C/staticLibrary/simpleCalcLib/calc.c
@@ -10,7 +10,7 @@ int multiply(int a, int b){
   return a*b;
 }
 
-float divide(float a, float b){
+int divide(int a, int b){
   if(!b) return -1;
   return a/b;
 }
